+++ src/components/playlist/index.ts
@@ -1,5 +1,5 @@
 import Playlist from './Playlist';
 import PlaylistButton from './PlaylistButton';
-import { usePlaylist } from './usePlaylist';
+import { usePlaylist } from './hooks/usePlaylist';
 
 export { Playlist, PlaylistButton, usePlaylist };
