+++ src/components/playlist/store/stores.ts
@@ -1,0 +1,9 @@
+import { atom } from 'jotai';
+
+type PlaylistStoreType = {
+  isOpenPlaylist: boolean;
+};
+
+export const playlistAtom = atom<PlaylistStoreType>({
+  isOpenPlaylist: false,
+});
